requirements.txt
newspaper3k
pydantic
wordfreq